proof of concept move of sklearn roc to plots
